pyproject.toml
[tool.poetry]
name = "python-study_60"
version = "0.1.0"
description = "for Python study #60"
authors = ["Kazuya Arai <kazuya.arai@uzabase.com>"]

[tool.poetry.dependencies]
python = "^3.8"
black = "^19.10b0"
logzero = "^1.5.0"
click = "^7.1.2"
lxml = "^4.5.2"
mecab-python3 = "^1.0.1"
wordcloud = "^1.7.0"
mypy = "^0.782"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
